commit 28.2 - 23/05/2024

+++ LISTA 2 - LING C/questao 8.c
@@ -7,10 +7,10 @@ int main(){
     puts("Digite o valor do produto: ");
     scanf("%f", & v_produto);
 
-    prestacao = (v_produto / 3);
+    prestacao = (int)(v_produto / 3);
     entrada = v_produto - (2 * prestacao);
 
-    printf("\nO valor da entrada será %.2f e o valor das duas prestações será %d\n", entrada, prestacao);
+    printf("\nO valor da entrada será %.2f e o valor das duas prestações será %.2f\n", entrada, prestacao);
 
     return 0;
 }
